Update to new tick processing
